Extract tovirtual, holdcall, unholdcall into call
